Fix dialyzer issue by not putting invalid data in `if_none_match` field.
Before this change, the `if_none_match` field, which should be a boolean,
was being used to store the value `consistent`. Now that the GPB
protobuf compiler is generating appropriately typed fields, dialyzer
became unhappy about that. This change uses the request context field
in the PB service `state` record to track the value of `consistent`
rather than stuffing it in a field that was meant for something else.

Added logic to reset the request context (`reset_state`) before returning
to the `riak_api_pb_server` process that called this service to make sure
it doesn't bleed into another request. If this service ever has some
sort of streaming added to it, this logic will have to be revisited
as some context may have to be retained across requests.