qry_buffers: avoid sorting accumulated chunks, merge new chunks instead
Previously, new chunks were being added to existing ones via
lists:sort(lists:append(Acc, Chunk)). Because Acc is always sorted, it
is possible to avoid the outer sort and use a less costly lists:merge
instead. Also, no sorting is done when the chunk ends up being written to
the backend.

With this optimisation, qbufs are slightly faster (100 vs 110 sec for
the fully in-mem comprehensive test), but it's not really convincing.