A new key epoch is only needed when coordinating
new_key_epoch was called for all puts where the local get returned a
notfound (no_old_object in the parlance of the recent refactor.) A new
epoch is only needed in certain situtions. This change means a new epoch
is only created when needed on no_old_object puts.

This is an intermediate step toward fixing the outstanding kv679 case
where a vnode has a stale/old/bad local vclock on disk. See the test
https://github.com/russelldb/riak_test/blob/rdb/gh-kv679-fb-byz/tests/kv679_dataloss_fb.erl
for more details.

This commit also simplifies the maybe_new_key_epoch function.