Hack the scheduling of AAE work to reduce initial window of broken AAE
When starting a new cluster in Riak 2.2.3, all the AAE stores are started in leagcy mode - as it is only once the ring gossipping has settled that the cluster cna determine that it is safe to upgrade to non-legacy AAE.

This then creates a problem, as each upgrade consumes a token, and by default ther eis only one token per hour to upgrade.  Asloe th eupgrade only wipes out the AAE store it does not rebuild it.

Once eacn node has 2 or 3 vnodes which have been wiped by the need to upgrade, then AAE exchanges grind to a halt - as both ends of the exchange need to be built and the same version.  But to become built, we first have to wait for all the vnodes to be wiped by an upgrade (an hour per vnode), and then each one to be built (an hour per vnode).  This means that on a typical cluster, AAE is broken for most of the first day of its life.

This hack tries to mitigate this:

1 - It doesn't spend a token on an upgrade (as an upgrade is not an expensive operation)
2 - It reorders the decision to that it will chose to upgrad ebefore it choses to rebuild (as otherwise there cna become a situation whereby a hashtree will rebuild one token period, and then clear itself to upgrade the next token period).

There's alsosome changes to make fold_heads safe - as  the legacy hash won't work correctly with the proxy_object.