WIP: a statem for the new HEAD->UPDATE code in get fsm
Needs work. The thing I'm stuck on is how to provide a correctly
ordered set of replies to the FSM such that I can terminate the FSM in
the next state funtion. Having access to the dynamic state, not
symbolic would help. I've hit a wall/or put myself in a corner/ so I'm
parking it and doing something easier for now.